Added /register route and added readWriteUsers db connection.
