+++ routes/index.js
@@ -1,6 +1,5 @@
 var express = require('express');
 var router = express.Router();
-var mysql = require('mysql');
 
 /* GET home page. */
 router.get('/', function(req, res, next) {
